Staging

+++ db/schema.rb
@@ -47,7 +47,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_07_08_150302) do
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
-    t.integer "blob_id", null: false
+    t.bigint "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
